# docs/swagger.yaml
basePath: /
definitions:
...
  models.Contact:
    description: Modèle complet d'une demande de contact
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      message:
        type: string
      status:
        $ref: '#/definitions/models.StatusType'
      subject:
        type: string
      submittedAt:
        type: string
    required:
    - email
    - firstName
    - lastName
    - message
    - subject
    type: object
  models.ContactCreate:
    description: modèle pour créer une demande de contact
    properties:
      email:
        example: jean.dupont@exemple.com
        type: string
      firstName:
        example: Jean
        type: string
      lastName:
        example: Dupont
        type: string
      message:
        example: J'aimerais avoir plus d'informations sur vos services.
        type: string
      subject:
        example: Demande d'information
        type: string
    required:
    - email
    - firstName
    - lastName
    - message
    - subject
    type: object
  models.ContactStatusUpdate:
    description: modèle pour mettre à jour le statut d'une demande de contact
    properties:
      status:
        allOf:
        - $ref: '#/definitions/models.StatusType'
        example: processing
    required:
    - status
    type: object
...
  models.StatusType:
    enum:
    - open
    - processing
    - closed
    - rejected
    type: string
    x-enum-varnames:
    - StatusOpen
    - StatusProcessing
    - StatusClosed
    - StatusRejected
...
host: localhost:8090
info:
  contact: {}
  description: API pour le projet PEC2 Backend
  title: API PEC2 Backend
  version: "1.0"
paths:
...
  /contact:
    post:
      consumes:
      - application/json
      description: Submit a new contact request with the provided information
      parameters:
      - description: Contact information
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactCreate'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Contact request submitted successfully, id: contact
            ID'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties: true
            type: object
      summary: Create a new contact request
      tags:
      - contacts
  /contacts:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all contact requests (Admin access only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Contact'
              type: array
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden - Admin access required'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all contact requests (Admin)
      tags:
      - contacts
  /contacts/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a contact request (Admin access only)
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update information
        in: body
        name: statusUpdate
        required: true
        schema:
          $ref: '#/definitions/models.ContactStatusUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Contact status updated successfully'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Contact not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update contact status (Admin)
      tags:
      - contacts
...
  /posts/{id}/comments:
    post:
      consumes:
      - application/json
      description: Create a new comment and broadcast it via SSE
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: JWT Token for web clients (optional)
        in: query
        name: token
        type: string
      - description: Comment content
        in: body
        name: comment
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Comment created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new comment for a post
      tags:
      - comments
  /posts/{id}/comments/sse:
    get:
      description: Connect to SSE to receive comments in real-time for a specific
        post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: JWT Token for web clients (optional)
        in: query
        name: token
        type: string
      responses:
        "200":
          description: Connected to SSE
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Invalid post ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Error setting up SSE'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Handle SSE connection for comments
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: 'Entrez le JWT avec le préfixe Bearer: Bearer <JWT>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"